connector/types: fix range returned by the `Type.UintRange` method

+++ connector/types/types.go
@@ -617,7 +617,7 @@ func (t Type) UIntRange() (min, max uint64) {
 		panic("type is not an UInt, UInt8, Int16, UInt24 or UInt64 type")
 	}
 	if t.pt != PtUInt64 {
-		return uint64(t.p), uint64(t.s)
+		return uint64(uint32(t.p)), uint64(uint32(t.s))
 	}
 	if i, ok := t.vl.(uintRange); ok {
 		return i.min, i.max
